+++ MekHQ/src/mekhq/campaign/mission/AbstractMissionTransition.java
@@ -86,6 +86,7 @@ import mekhq.campaign.personnel.ranks.RankSystem;
 import mekhq.campaign.personnel.ranks.RankValidator;
 import mekhq.campaign.personnel.ranks.Ranks;
 import mekhq.campaign.stratCon.StratConCampaignState;
+import mekhq.campaign.stratCon.StratConTrackState;
 import mekhq.campaign.stratCon.SupportPointNegotiation;
 import mekhq.campaign.unit.Unit;
 import mekhq.campaign.universe.Faction;
@@ -1414,6 +1415,56 @@ public class AbstractMissionTransition {
      */
     public int getRepairLocation() {
         return Unit.SITE_FACILITY_BASIC;
+    }
+
+
+    /**
+     * Calculates the number of required Victory Points (VP) needed to achieve overall success for this StratCon
+     * contract.
+     *
+     * <p>The calculation is based on several averaged campaign parameters:
+     * <ul>
+     *     <li><b>Base requirement</b> — Required number of combat teams multiplied by the contract length.</li>
+     *     <li><b>Scenario odds</b> — The mean scenario-odds percentage across all StratCon tracks, converted to a
+     *     probability.</li>
+     *     <li><b>Turning point chance</b> — A scaling factor based on command rights: {@code INTEGRATED} contracts
+     *     assume a 100% chance, while all others use a one-third chance.</li>
+     * </ul>
+     *
+     * <p>The final result estimates the expected number of Turning Points the player must win for overall contract
+     * success. If the player loses a handful of Turning Points, they should still be able to win the contract by
+     * being proactive in the Area of Operations.</p>
+     *
+     * @return the required number of Victory Points, rounded up to the nearest integer
+     *
+     * @author Illiani
+     * @since 0.50.10
+     */
+    public int getRequiredVictoryPoints() {
+        if (getStratConCampaignState() == null) {
+            return 0;
+        }
+
+        double baseRequirement = getRequiredCombatTeams();
+
+        int duration = getLengthInMonths();
+        if (getContractType().isGarrisonType()) {
+            duration = (int) ceil(duration * 0.75); // We assume around 25% of the contract will be peaceful
+        }
+
+        double trackCount = 0;
+        int totalScenarioOdds = 0;
+        for (StratConTrackState trackState : getStratConCampaignState().getTracks()) {
+            trackCount++;
+            totalScenarioOdds += trackState.getScenarioOdds();
+        }
+
+        double meanScenarioOdds = totalScenarioOdds / trackCount;
+        double scenarioOdds = meanScenarioOdds / 100.0;
+        double turningPointChance = (getCommandRights() == ContractCommandRights.INTEGRATED ? 1.0 : 0.33);
+
+        // This result gives us the average number of Turning Points expected for the contract
+        return (int) ceil(baseRequirement * duration * scenarioOdds * turningPointChance);
     }
 
     /**
